refactoring explore page
